PycharmProjects/chatbot/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
line-bot-sdk==1.19.0